[S12P21E206-208] 실시간 리얼타임 기능 구현

+++ back/notification/src/main/java/com/notification/controller/OrderNotificationController.java
@@ -16,7 +16,7 @@ public class OrderNotificationController {
     OrderService orderService;
 
     @PostMapping("/total-order-amount")
-    public ResponseEntity<?> sendToAllTotalOrderAmount(@RequestBody int totalOrderAmount) {
+    public ResponseEntity<?> sendToAllTotalOrderAmount(@RequestBody Long totalOrderAmount) {
         orderService.sendToAllTotalOrderAmount(totalOrderAmount);
         return new ResponseEntity<>(Response.create(SEND_TO_ALL_TOTAL_ORDER_AMOUNT, null), SEND_TO_ALL_TOTAL_ORDER_AMOUNT.getHttpStatus());
     }
