+++ back/notification/src/main/java/com/notification/service/OrderService.java
@@ -12,7 +12,7 @@ public class OrderService {
         this.simpMessagingTemplate = simpMessagingTemplate;
     }
 
-    public void sendToAllTotalOrderAmount(int totalOrderAmount) {
+    public void sendToAllTotalOrderAmount(Long totalOrderAmount) {
         simpMessagingTemplate.convertAndSend("/topic/totalAmount", totalOrderAmount);
     }
